worker/requirements.txt
google-cloud-pubsub==2.21.0
